requirements.txt
orion-ml
openai
pandas
numpy
PyGithub # for document collection